scripts/requirements.txt
bottle==0.12.21
pynput==1.7.6
pyserial==3.5
requests==2.28.0
